fix upgrade mo_user in 2.0-dev
